chore(perps): remove perps toast provider from perps view
