+++ ui/components/app/perps/perps-toast/perps-toast-provider.tsx
@@ -7,40 +7,23 @@ import React, {
   useState,
   type ReactNode,
 } from 'react';
-import {
-  AvatarIcon,
-  AvatarIconSeverity,
-  AvatarIconSize,
-  Box,
-  Icon,
-  IconColor,
-  IconName,
-  IconSize,
-} from '@metamask/design-system-react';
+import { Box } from '@metamask/design-system-react';
 import { Toast } from '../../../multichain/toast';
 import { useI18nContext } from '../../../../hooks/useI18nContext';
+import {
+  PERPS_TOAST_KEYS,
+  type PerpsToastKey,
+  type PerpsToastVariant,
+} from './perps-toast.constants';
+import {
+  getPerpsToastIcon,
+  PERPS_TOAST_PRESENTATION_BY_KEY,
+  PERPS_TOAST_PRESENTATION_BY_VARIANT,
+  type PerpsToastPresentation,
+} from './perps-toast.presentation';
 
-export type PerpsToastVariant = 'info' | 'success' | 'error';
-
-export const PERPS_TOAST_KEYS = {
-  CLOSE_FAILED: 'perpsToastCloseFailed',
-  CLOSE_IN_PROGRESS: 'perpsToastCloseInProgress',
-  MARGIN_ADD_SUCCESS: 'perpsToastMarginAddSuccess',
-  MARGIN_ADJUSTMENT_FAILED: 'perpsToastMarginAdjustmentFailed',
-  MARGIN_REMOVE_SUCCESS: 'perpsToastMarginRemoveSuccess',
-  ORDER_FAILED: 'perpsToastOrderFailed',
-  ORDER_FILLED: 'perpsToastOrderFilled',
-  ORDER_PLACED: 'perpsToastOrderPlaced',
-  ORDER_SUBMITTED: 'perpsToastOrderSubmitted',
-  SUBMIT_IN_PROGRESS: 'perpsToastSubmitInProgress',
-  TRADE_SUCCESS: 'perpsToastTradeSuccess',
-  UPDATE_FAILED: 'perpsToastUpdateFailed',
-  UPDATE_IN_PROGRESS: 'perpsToastUpdateInProgress',
-  UPDATE_SUCCESS: 'perpsToastUpdateSuccess',
-} as const;
-
-export type PerpsToastKey =
-  (typeof PERPS_TOAST_KEYS)[keyof typeof PERPS_TOAST_KEYS];
+export { PERPS_TOAST_KEYS };
+export type { PerpsToastKey, PerpsToastVariant };
 
 export type PerpsToastRouteState = {
   perpsToastKey?: PerpsToastKey;
@@ -101,129 +84,6 @@ const getDefaultAutoHideTime = (
   }
 
   return undefined;
-};
-
-type PerpsToastIconConfig =
-  | {
-      color: IconColor;
-      dataTestId: string;
-      name: IconName;
-      type: 'spinner';
-    }
-  | {
-      className?: string;
-      dataTestId: string;
-      name: IconName;
-      severity: AvatarIconSeverity;
-      size: AvatarIconSize;
-      type: 'avatar-icon';
-    };
-
-type PerpsToastPresentation = {
-  icon: PerpsToastIconConfig;
-  variant: PerpsToastVariant;
-};
-
-const PERPS_TOAST_PRESENTATION_BY_VARIANT: Record<
-  PerpsToastVariant,
-  PerpsToastPresentation
-> = {
-  error: {
-    variant: 'error',
-    icon: {
-      type: 'avatar-icon',
-      name: IconName.Warning,
-      severity: AvatarIconSeverity.Error,
-      size: AvatarIconSize.Md,
-      dataTestId: 'perps-toast-icon-warning',
-    },
-  },
-  info: {
-    variant: 'info',
-    icon: {
-      type: 'spinner',
-      name: IconName.Loading,
-      color: IconColor.PrimaryDefault,
-      dataTestId: 'perps-toast-icon-loading',
-    },
-  },
-  success: {
-    variant: 'success',
-    icon: {
-      type: 'avatar-icon',
-      name: IconName.CheckBold,
-      severity: AvatarIconSeverity.Success,
-      className: 'perps-toast__success-icon',
-      size: AvatarIconSize.Md,
-      dataTestId: 'perps-toast-icon-check-bold',
-    },
-  },
-};
-
-// Shared avatar-style error presentation for perps failed toasts.
-const PERPS_ERROR_PRESENTATION: PerpsToastPresentation = {
-  variant: 'error',
-  icon: {
-    type: 'avatar-icon',
-    name: IconName.Warning,
-    severity: AvatarIconSeverity.Error,
-    size: AvatarIconSize.Md,
-    dataTestId: 'perps-toast-icon-warning',
-  },
-};
-
-const PERPS_TOAST_PRESENTATION_BY_KEY: Record<
-  PerpsToastKey,
-  PerpsToastPresentation
-> = {
-  [PERPS_TOAST_KEYS.CLOSE_FAILED]: PERPS_ERROR_PRESENTATION,
-  [PERPS_TOAST_KEYS.CLOSE_IN_PROGRESS]:
-    PERPS_TOAST_PRESENTATION_BY_VARIANT.info,
-  [PERPS_TOAST_KEYS.MARGIN_ADD_SUCCESS]:
-    PERPS_TOAST_PRESENTATION_BY_VARIANT.success,
-  [PERPS_TOAST_KEYS.MARGIN_ADJUSTMENT_FAILED]: PERPS_ERROR_PRESENTATION,
-  [PERPS_TOAST_KEYS.MARGIN_REMOVE_SUCCESS]:
-    PERPS_TOAST_PRESENTATION_BY_VARIANT.success,
-  [PERPS_TOAST_KEYS.ORDER_FAILED]: PERPS_ERROR_PRESENTATION,
-  [PERPS_TOAST_KEYS.ORDER_FILLED]: PERPS_TOAST_PRESENTATION_BY_VARIANT.success,
-  [PERPS_TOAST_KEYS.ORDER_PLACED]: PERPS_TOAST_PRESENTATION_BY_VARIANT.success,
-  [PERPS_TOAST_KEYS.ORDER_SUBMITTED]: PERPS_TOAST_PRESENTATION_BY_VARIANT.info,
-  [PERPS_TOAST_KEYS.SUBMIT_IN_PROGRESS]:
-    PERPS_TOAST_PRESENTATION_BY_VARIANT.info,
-  [PERPS_TOAST_KEYS.TRADE_SUCCESS]: PERPS_TOAST_PRESENTATION_BY_VARIANT.success,
-  [PERPS_TOAST_KEYS.UPDATE_FAILED]: PERPS_ERROR_PRESENTATION,
-  [PERPS_TOAST_KEYS.UPDATE_IN_PROGRESS]:
-    PERPS_TOAST_PRESENTATION_BY_VARIANT.info,
-  [PERPS_TOAST_KEYS.UPDATE_SUCCESS]:
-    PERPS_TOAST_PRESENTATION_BY_VARIANT.success,
-};
-
-const getToastIcon = ({ icon }: PerpsToastPresentation): ReactNode => {
-  if (icon.type === 'spinner') {
-    return (
-      <Box
-        className="flex h-8 w-8 items-center justify-center"
-        data-testid={icon.dataTestId}
-      >
-        <Icon
-          name={icon.name}
-          size={IconSize.Xl}
-          color={icon.color}
-          className="animate-spin"
-        />
-      </Box>
-    );
-  }
-
-  return (
-    <AvatarIcon
-      iconName={icon.name}
-      size={icon.size}
-      severity={icon.severity}
-      className={icon.className}
-      data-testid={icon.dataTestId}
-    />
-  );
 };
 
 type PerpsToastProviderProps = {
@@ -295,7 +155,7 @@ export const PerpsToastProvider = ({ children }: PerpsToastProviderProps) => {
         <Box className="toasts-container bottom-20 w-[calc(100%-32px)] max-w-[408px]">
           <Toast
             key={activeToast.id}
-            startAdornment={getToastIcon(activeToast.presentation)}
+            startAdornment={getPerpsToastIcon(activeToast.presentation)}
             text={activeToast.message}
             description={activeToast.description}
             className="perps-toast"
