+++ ui/components/app/perps/perps-view.tsx
@@ -31,7 +31,6 @@ import {
 import { PerpsSupportLearn } from './perps-support-learn';
 import { PerpsTutorialModal } from './perps-tutorial-modal';
 import { PerpsWatchlist } from './perps-watchlist';
-import { PerpsToastProvider } from './perps-toast';
 
 /**
  * PerpsView component displays the perpetuals trading view
@@ -181,58 +180,56 @@ export const PerpsView: React.FC = () => {
   }
 
   return (
-    <PerpsToastProvider>
-      <Box
-        flexDirection={BoxFlexDirection.Column}
-        gap={4}
-        data-testid="perps-view"
-      >
-        {/* Balance header with Add funds / Withdraw dropdown */}
-        <PerpsBalanceDropdown
-          hasPositions={hasPositions}
-          onAddFunds={triggerDeposit}
-          onWithdraw={triggerWithdraw}
-        />
+    <Box
+      flexDirection={BoxFlexDirection.Column}
+      gap={4}
+      data-testid="perps-view"
+    >
+      {/* Balance header with Add funds / Withdraw dropdown */}
+      <PerpsBalanceDropdown
+        hasPositions={hasPositions}
+        onAddFunds={triggerDeposit}
+        onWithdraw={triggerWithdraw}
+      />
 
-        {/* Positions + Orders sections */}
-        {batchActionError ? (
-          <Text variant={TextVariant.BodySm} color={TextColor.ErrorDefault}>
-            {batchActionError}
-          </Text>
-        ) : null}
+      {/* Positions + Orders sections */}
+      {batchActionError ? (
+        <Text variant={TextVariant.BodySm} color={TextColor.ErrorDefault}>
+          {batchActionError}
+        </Text>
+      ) : null}
 
-        <PerpsPositionsOrders
-          positions={positions}
-          orders={orders}
-          onCloseAllPositions={handleCloseAllPositions}
-          onCancelAllOrders={handleCancelAllOrders}
-          isCloseAllPending={isCloseAllPending}
-          isCancelAllPending={isCancelAllPending}
-        />
+      <PerpsPositionsOrders
+        positions={positions}
+        orders={orders}
+        onCloseAllPositions={handleCloseAllPositions}
+        onCancelAllOrders={handleCancelAllOrders}
+        isCloseAllPending={isCloseAllPending}
+        isCancelAllPending={isCancelAllPending}
+      />
 
-        {/* Watchlist */}
-        <PerpsWatchlist />
+      {/* Watchlist */}
+      <PerpsWatchlist />
 
-        {/* Explore markets */}
-        <PerpsExploreMarkets
-          cryptoMarkets={cryptoMarkets}
-          hip3Markets={hip3Markets}
-        />
+      {/* Explore markets */}
+      <PerpsExploreMarkets
+        cryptoMarkets={cryptoMarkets}
+        hip3Markets={hip3Markets}
+      />
 
-        {/* Recent Activity */}
-        <PerpsRecentActivity
-          transactions={recentActivityTransactions}
-          maxTransactions={PERPS_RECENT_ACTIVITY_MAX_TRANSACTIONS}
-          isLoading={recentActivityLoading}
-          error={recentActivityError}
-        />
+      {/* Recent Activity */}
+      <PerpsRecentActivity
+        transactions={recentActivityTransactions}
+        maxTransactions={PERPS_RECENT_ACTIVITY_MAX_TRANSACTIONS}
+        isLoading={recentActivityLoading}
+        error={recentActivityError}
+      />
 
-        {/* Support & Learn */}
-        <PerpsSupportLearn />
-        {/* Tutorial Modal */}
-        <PerpsTutorialModal />
-      </Box>
-    </PerpsToastProvider>
+      {/* Support & Learn */}
+      <PerpsSupportLearn />
+      {/* Tutorial Modal */}
+      <PerpsTutorialModal />
+    </Box>
   );
 };
