+++ ui/components/multichain/account-overview/account-overview-tabs.tsx
@@ -30,7 +30,11 @@ import {
 import AssetList from '../../app/assets/asset-list';
 import DeFiTab from '../../app/assets/defi-list/defi-tab';
 import NftsTab from '../../app/assets/nfts/nfts-tab';
-import { PerpsView, PerpsViewStreamBoundary } from '../../app/perps';
+import {
+  PerpsToastProvider,
+  PerpsView,
+  PerpsViewStreamBoundary,
+} from '../../app/perps';
 import { Tab, Tabs } from '../../ui/tabs';
 import { useTokenBalances } from '../../../hooks/useTokenBalances';
 import { ActivityList } from '../activity-v2/activity-list';
@@ -181,11 +185,13 @@ export const AccountOverviewTabs = ({
             tabKey={AccountOverviewTabKey.Perps}
             data-testid="account-overview__perps-tab"
           >
-            <ErrorBoundary key="perps">
-              <PerpsViewStreamBoundary>
-                <PerpsView />
-              </PerpsViewStreamBoundary>
-            </ErrorBoundary>
+            <PerpsToastProvider>
+              <ErrorBoundary key="perps">
+                <PerpsViewStreamBoundary>
+                  <PerpsView />
+                </PerpsViewStreamBoundary>
+              </ErrorBoundary>
+            </PerpsToastProvider>
           </Tab>
         )}
